add_fcs

add_fcs

Usage

add_fcs(
  df,
  cutoffs = "normal 21.5-35",
  fcs_cereal,
  fcs_legumes,
  fcs_veg,
  fcs_fruit,
  fcs_meat,
  fcs_dairy,
  fcs_sugar,
  fcs_oil
)

Arguments

df

the clean dataset

cutoffs

either "normal 21.5-35", or "alternative 28-42". The default is set to normal

fcs_cereal

the name of the variable that indicates the number of days cereals were consumed

fcs_legumes

the name of the variable that indicates the number of days legumes were consumed

fcs_veg

the name of the variable that indicates the number of days vegetables were consumed

fcs_fruit

the name of the variable that indicates the number of days fruits were consumed

fcs_meat

the name of the variable that indicates the number of days meat/fish were consumed

fcs_dairy

the name of the variable that indicates the number of days dairy were consumed

fcs_sugar

the name of the variable that indicates the number of days cereals was consumed

fcs_oil

the name of the variable that indicates the number of days oild were consumed

Value

the dataset with fcs_score and fcs_cat computed, as well as the 8 weighted food groups

Examples

df1 <- data.frame(
  cereal = c(1, 2, 3, 2, 5, 6, 7),
  pulses = c(3, 4, 5, 6, 1, 6, 5),
  vegetables = c(3, 2, 1, 6, 5, 4, 3),
  fruits = c(1, 4, 6, 2, 2, 2, 4),
  meat = c(5, 4, 3, 2, 7, 4, 5),
  milk = c(1, 2, 6, 7, 3, 4, 2),
  sugar = c(1, 7, 6, 5, 2, 3, 4),
  oil = c(2, 3, 6, 5, 1, 7, 4)
)
add_fcs(df1,
  cutoffs = "normal 21.5-35", fcs_cereal = "cereal", fcs_legumes = "pulses", fcs_veg = "vegetables",
  fcs_fruit = "fruits", fcs_meat = "meat", fcs_dairy = "milk", fcs_sugar = "sugar", fcs_oil = "oil"
)
#>   cereal pulses vegetables fruits meat milk sugar oil fcs_weight_cereal1
#> 1      1      3          3      1    5    1     1   2                  2
#> 2      2      4          2      4    4    2     7   3                  4
#> 3      3      5          1      6    3    6     6   6                  6
#> 4      2      6          6      2    2    7     5   5                  4
#> 5      5      1          5      2    7    3     2   1                 10
#> 6      6      6          4      2    4    4     3   7                 12
#> 7      7      5          3      4    5    2     4   4                 14
#>   fcs_weight_legume2 fcs_weight_dairy3 fcs_weight_meat4 fcs_weight_veg5
#> 1                  9                 4               20               3
#> 2                 12                 8               16               2
#> 3                 15                24               12               1
#> 4                 18                28                8               6
#> 5                  3                12               28               5
#> 6                 18                16               16               4
#> 7                 15                 8               20               3
#>   fcs_weight_fruit6 fcs_weight_oil7 fcs_weight_sugar8 fcs_score    fcs_cat
#> 1                 1             1.0               0.5      40.5 Acceptable
#> 2                 4             1.5               3.5      51.0 Acceptable
#> 3                 6             3.0               3.0      70.0 Acceptable
#> 4                 2             2.5               2.5      71.0 Acceptable
#> 5                 2             0.5               1.0      61.5 Acceptable
#> 6                 2             3.5               1.5      73.0 Acceptable
#> 7                 4             2.0               2.0      68.0 Acceptable